Create A3 Code Upcycle clothing
This is a basic HTML for my website Upcycle clothing. HTML is key for building a website, I improved on my A2 by adding a bit of CSS to add some design structure. Obviously basic but is the bones for a website and in turn has allowed me to improve on my code literacy